refactorizando aparte del código
