+++ src/css/index.css
@@ -1,3 +1,2 @@
 @import "./auth.css";
-@import "./desktop.css";
 @import "./share.css";
